+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 ## 2026-08-08
 
+- **@poc-plattform-kit/api** `0.19.0` → `0.20.0` (minor)
+- **@poc-plattform-kit/web** `0.14.0` → `0.14.1` (patch)
+- **@poc-plattform-kit/db** `0.1.0` → `0.1.1` (patch)
+## 2026-08-08
+
+- **@poc-plattform-kit/api** `0.18.0` → `0.19.0` (minor)
+- **@poc-plattform-kit/web** `0.13.0` → `0.14.0` (minor)
+- **@poc-plattform-kit/api-client** `0.10.0` → `0.11.0` (minor)
+- **@poc-plattform-kit/pillar-tenant** `0.5.0` → `0.6.0` (minor)
+## 2026-08-08
+
+- **@poc-plattform-kit/api** `0.17.0` → `0.18.0` (minor)
+- **@poc-plattform-kit/web** `0.12.0` → `0.13.0` (minor)
+- **@poc-plattform-kit/api-client** `0.9.0` → `0.10.0` (minor)
+- **@poc-plattform-kit/pillar-permissions** `0.3.0` → `0.3.1` (patch)
+- **@poc-plattform-kit/pillar-tenant** `0.4.0` → `0.5.0` (minor)
+## 2026-08-08
+
+- **@poc-plattform-kit/api** `0.16.0` → `0.17.0` (minor)
+## 2026-08-08
+
 - **@poc-plattform-kit/api** `0.15.0` → `0.16.0` (minor)
 - **@poc-plattform-kit/web** `0.11.0` → `0.12.0` (minor)
 - **@poc-plattform-kit/api-client** `0.8.0` → `0.9.0` (minor)